refactor: move Sharadar provider to provider/sharadar sub-package
