+++ cmd/import.go
@@ -9,6 +9,7 @@ import (
 	"github.com/penny-vault/pvdata/data"
 	"github.com/penny-vault/pvdata/library"
 	"github.com/penny-vault/pvdata/provider"
+	"github.com/penny-vault/pvdata/provider/sharadar"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
@@ -67,7 +68,7 @@ Examples:
 				log.Fatal().Str("file", f).Msg("file does not exist")
 			}
 
-			detected, detectErr := provider.DetectSharadarDataset(f)
+			detected, detectErr := sharadar.DetectSharadarDataset(f)
 			if detectErr != nil {
 				log.Warn().Err(detectErr).Str("file", f).Msg("could not detect dataset from filename, proceeding anyway")
 			} else if detected != sub.Dataset {
